Refactor writer packages (#517)
