--- src/cmds/mod.rs
@@ -1,6 +1,0 @@
-pub mod core;
-pub mod func;
-pub mod info;
-pub mod preview;
-pub mod repo;
-pub mod shell;
