+++ src/lib.rs
@@ -6,7 +6,6 @@ extern crate anyhow;
 mod actor;
 mod cheatsh;
 mod clipboard;
-mod cmds;
 mod env_var;
 mod extractor;
 mod filesystem;
@@ -20,6 +19,7 @@ mod shell;
 mod structures;
 mod terminal;
 mod tldr;
+mod ui;
 mod url;
 mod welcome;
 mod writer;
